requirements.txt
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.3.0
click==8.1.8
cramjam==2.9.1
dnspython==2.7.0
ecdsa==0.19.0
fastapi==0.115.11
h11==0.14.0
idna==3.10
motor==3.7.0
pyasn1==0.4.8
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.11.2
python-jose==3.4.0
python-snappy==0.7.3
rsa==4.9
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
typing_extensions==4.12.2
uvicorn==0.34.0
